fix(features): remove unused code and move funcs to utils file
