+++ src/v3/features/ministry/add_service_time_modal_window/index.tsx
@@ -12,7 +12,6 @@ import { useContext, useEffect, useRef, useState } from 'react';
 import { CustomDropdownContainer, CustomDropdownItem, CustomDropdownMenu } from '@components/dropdown';
 import {
   IconAdd,
-  IconClose,
   IconError,
   IconLanguageCourse,
   IconPersonalDay,
@@ -24,6 +23,7 @@ import { EditAndAddBibleStudyContext } from '../EditAndAddBibleStudyContext';
 import CustomButton from '@components/button';
 import CustomDatePicker from '@components/date_picker';
 import CustomInfoMessage from '@components/info-message';
+import { convertDurationInSecondsToString, convertDurationStringToSeconds } from '../utils';
 
 /**
  * Add Service Time Modal Window component.
@@ -35,35 +35,6 @@ export const AddServiceTimeModalWindow = (props: AddServiceTimeModalWindowProps)
   const { editAndAddBibleStudyData, setEditAndAddBibleStudyData } = useContext(EditAndAddBibleStudyContext);
 
   const { t } = useAppTranslation();
-
-  /**
-   * Converts a duration string in the format 'HH:MM' to seconds.
-   * @param {string} duration - The duration string to convert.
-   * @returns {number} The duration in seconds.
-   */
-  const convertDurationStringToSeconds = (duration: string): number => {
-    const [hours, minutes] = duration.split(':');
-    return parseInt(hours) * 3600 + parseInt(minutes) * 60;
-  };
-
-  /**
-   * Converts a duration in seconds to a string format 'HH:MM'.
-   * @param {number} seconds - The duration in seconds to convert.
-   * @returns {string} The duration string in 'HH:MM' format.
-   */
-  const convertDurationInSecondsToString = (seconds: number): string => {
-    const hours = Math.floor(seconds / 3600);
-    const minutes = Math.floor((seconds % 3600) / 60);
-
-    if (hours === 0 && minutes === 0) {
-      return '00:00';
-    }
-
-    const formattedHours = hours < 10 ? '0' + hours.toString() : hours.toString();
-    const formattedMinutes = minutes < 10 ? '0' + minutes.toString() : minutes.toString();
-
-    return `${formattedHours}:${formattedMinutes}`;
-  };
 
   const [localDurationInSeconds, setLocalDurationInSeconds] = useState(0);
